[MERGE] Add support for commercial building models driven by CEUS data (#72)
* Added initial implementation of CEUS data

* Working update

* Continued work on CEUS commercial loads

* Finished power calcs

* Fixed problem adding enduses

* More cleanup. Added stubs for price and temperature sensitivity calcs.

* Fixed problem finding enduses

* Working update

* Updated to more compact CEUS data format (see ceus_data repo on github)

* Added all building types and enduses in zone 1 to CEUS test file

* Removing unneeded test file

* Merge branch 'master' into dchassin/ceus

Include readying branch for merge into master

* Added temperature, solar, and price sensitivity

* Removed PNG files

* Minor fixes to CEUS model

* Moved sensivitities from object level to enduse level

Added output from main meter. Fixed meter so it resets accumulators before presyncs are done.

* Fixed warnings in commercial/ceus.{cpp,h}

* Fixed warning in gldcore/complex.h

* Fix of wrong integer size in enumeration conversion (#188)

* Update ceus.cpp

* Fixed inverter error

* Fixed problem with meter accumulators

* Added matplotlib to centos build to support CEUS test

* Fixed linux centos warnings from last changes

+++ generators/inverter.cpp
@@ -574,9 +574,8 @@ int inverter::init(OBJECT *parent)
 	OBJECT *objBattery = NULL;
 	size_t index = 0;
 
-	if ( parent != NULL )
-	{
-		if ( (parent->flags & OF_INIT) != OF_INIT )
+	if(parent != NULL){
+		if((parent->flags & OF_INIT) != OF_INIT)
 		{
 			verbose("init() deferring initialization");
 			return 2; // defer
@@ -615,7 +614,7 @@ int inverter::init(OBJECT *parent)
 		};
 		/// @todo use triplex property mapping instead of assuming memory order for meter variables (residential, low priority) (ticket #139)
 	
-		for (i=0; i<sizeof(map)/sizeof(map[0]); i++)
+		for ( i = 0; i < sizeof(map)/sizeof(map[0]) ; i++ )
 			*(map[i].var) = get_complex(parent,map[i].varname);
 
 		//Map status
@@ -673,7 +672,7 @@ int inverter::init(OBJECT *parent)
 		};
 
 		// attach meter variables to each circuit
-		for (i=0; i<sizeof(map)/sizeof(map[0]); i++)
+		for ( i = 0 ; i < sizeof(map)/sizeof(map[0]) ; i++ )
 		{
 			if ((*(map[i].var) = get_complex(parent,map[i].varname))==NULL)
 			{
@@ -933,7 +932,8 @@ int inverter::init(OBJECT *parent)
 					//std::string tempV = "";
 					tempV = "";
 					tempQ = "";
-					for(size_t i = 0; i < VoltVArSchedInput.length(); i++)	{
+					for( size_t i = 0; i < VoltVArSchedInput.length() ; i++ )	
+					{
 						if(VoltVArSchedInput[i] != ',')	{
 							if(cntr % 2 == 0)
 								tempV += VoltVArSchedInput[i];
@@ -970,7 +970,8 @@ int inverter::init(OBJECT *parent)
 					int cntr = 0;
 					tempf = "";
 					tempP = "";
-					for(size_t i = 0; i < freq_pwrSchedInput.length(); i++)	{
+					for( size_t i = 0 ; i < freq_pwrSchedInput.length() ; i++ )	
+					{
 						if(freq_pwrSchedInput[i] != ',')	{
 							if(cntr % 2 == 0)
 								tempf += freq_pwrSchedInput[i];
@@ -2886,7 +2887,7 @@ TIMESTAMP inverter::sync(TIMESTAMP t0, TIMESTAMP t1)
 		else	//FOUR_QUADRANT code
 		{
 			//FOUR_QUADRANT model (originally written for NAS/CES, altered for PV)
-			double VA_Efficiency, temp_PF, temp_QVal, P_in = 0.0, net_eff = 0.0; //Ab added last two
+			double VA_Efficiency, temp_PF, temp_QVal, P_in = 0, net_eff = 0; //Ab added last two
 			complex temp_VA;
 			complex battery_power_out = complex(0,0);
 			if (four_quadrant_control_mode != FQM_VOLT_VAR) {
@@ -3820,7 +3821,7 @@ TIMESTAMP inverter::postsync(TIMESTAMP t0, TIMESTAMP t1)
 	TIMESTAMP t2 = TS_NEVER;		//By default, we're done forever!
 	LOAD_FOLLOW_STATUS new_lf_status = IDLE;
 	PF_REG_STATUS new_pf_reg_status = PFRS_UNKNOWN;
-	double new_lf_dispatch_power = 0.0, curr_power_val, diff_power_val;				
+	double new_lf_dispatch_power = 0, curr_power_val, diff_power_val;				
 	double new_pf_reg_distpatch_VAR = 0.0, curr_real_power_val, curr_reactive_power_val, curr_pf, Q_out, Q_available;
 	double scaling_factor, Q_target;
 	complex temp_current_val[3];
@@ -4891,8 +4892,8 @@ SIMULATIONMODE inverter::inter_deltaupdate(unsigned int64 delta_time, unsigned l
 	double prev_error_ed;
 	double prev_error_eq;
 	bool ramp_change;
-	int i = 0;
-	double ieee_1547_double = 0.0;
+	size_t i = 0;
+	double ieee_1547_double = 0;
 	complex temp_current_val[3];
 	complex power_val[3];
 
@@ -6167,6 +6168,9 @@ SIMULATIONMODE inverter::inter_deltaupdate(unsigned int64 delta_time, unsigned l
 							pred_state.P_Out[i] = (pCircuit_V[i] * ~(I_Out[i])).Re();
 							pred_state.Q_Out[i] = (pCircuit_V[i] * ~(I_Out[i])).Im();
 
+							// if (Pref > 0) {
+							// 	int stop_temp = 0;
+							// }
 							if (pCircuit_V[i].Mag() > 0.0)
 							{
 								pred_state.ed[i] = ((~(complex(Pref/3.0, Qref_PI[i])/(pCircuit_V[i]))) - (~(complex(pred_state.P_Out[i],pred_state.Q_Out[i])/(pCircuit_V[i])))).Re();
